refactor: break parser apart
