+++ src/parser/items.rs
@@ -1,0 +1,512 @@
+use crate::{
+    grammar::{
+        Attributes, BitflagsDefItem, BitflagsDefinition, BitflagsStatement, EnumDefItem,
+        EnumDefinition, EnumStatement, ItemDefinition, ItemDefinitionInner, TypeDefItem,
+        TypeDefinition, TypeField, TypeStatement,
+    },
+    span::{ItemLocation, Span},
+    tokenizer::TokenKind,
+};
+
+use super::{ParseError, core::Parser};
+
+impl Parser {
+    pub(crate) fn parse_item_definition(&mut self) -> Result<ItemDefinition, ParseError> {
+        // Capture the start position
+        let start_pos = self.current().location.span.start;
+
+        let mut doc_comments = self.collect_doc_comments();
+        let attributes = if matches!(self.peek(), TokenKind::Hash) {
+            self.parse_attributes()?
+        } else {
+            Attributes::default()
+        };
+
+        // Remember the line where attributes ended (or where we currently are if no attributes)
+        // We need to check this before we start collecting comments
+        let attributes_end_line = if !attributes.0.is_empty() && self.pos > 0 {
+            // Get the line from the previous token (the ] that closed the attributes)
+            self.tokens[self.pos - 1].location.span.end.line
+        } else {
+            // No attributes, so comments can't be inline with them
+            0 // Use 0 as a sentinel value that won't match any real line
+        };
+
+        // Collect comments after attributes, separating inline from following
+        let mut inline_trailing_comments = Vec::new();
+        let mut following_comments = Vec::new();
+        while matches!(
+            self.peek(),
+            TokenKind::Comment(_) | TokenKind::MultiLineComment(_)
+        ) {
+            let comment_line = self.current().location.span.start.line;
+            if let Some(comment) = self.collect_comment() {
+                if comment_line == attributes_end_line {
+                    // Comment is on the same line as the attributes
+                    inline_trailing_comments.push(comment);
+                } else {
+                    // Comment is on a following line
+                    following_comments.push(comment);
+                }
+            }
+        }
+
+        // Also collect doc comments that appear after attributes
+        let after_attr_doc_comments = self.collect_doc_comments();
+        doc_comments.extend(after_attr_doc_comments);
+
+        let visibility = self.parse_visibility()?;
+
+        match self.peek() {
+            TokenKind::Type => {
+                self.advance();
+                let (name, _) = self.expect_ident()?;
+                let mut def = TypeDefinition {
+                    items: Vec::new(),
+                    attributes,
+                    inline_trailing_comments: inline_trailing_comments.clone(),
+                    following_comments: following_comments.clone(),
+                };
+
+                // Support both "type Name;" and "type Name { ... }"
+                if matches!(self.peek(), TokenKind::Semi) {
+                    self.advance(); // Consume semicolon
+                } else {
+                    self.expect(TokenKind::LBrace)?;
+                    def.items = self.parse_type_def_items()?;
+                    self.expect(TokenKind::RBrace)?;
+                }
+
+                // Capture the end position
+                let end_pos = if self.pos > 0 {
+                    self.tokens[self.pos - 1].location.span.end
+                } else {
+                    self.current().location.span.end
+                };
+
+                Ok(ItemDefinition {
+                    visibility,
+                    name,
+                    doc_comments,
+                    inner: ItemDefinitionInner::Type(def),
+                    location: ItemLocation::new(
+                        self.filename.clone(),
+                        Span::new(start_pos, end_pos),
+                    ),
+                })
+            }
+            TokenKind::Enum => {
+                self.advance();
+                let (name, _) = self.expect_ident()?;
+                self.expect(TokenKind::Colon)?;
+                let type_ = self.parse_type()?;
+                self.expect(TokenKind::LBrace)?;
+                let items = self.parse_enum_def_items()?;
+                self.expect(TokenKind::RBrace)?;
+
+                // Capture the end position
+                let end_pos = if self.pos > 0 {
+                    self.tokens[self.pos - 1].location.span.end
+                } else {
+                    self.current().location.span.end
+                };
+
+                Ok(ItemDefinition {
+                    visibility,
+                    name,
+                    doc_comments,
+                    inner: ItemDefinitionInner::Enum(EnumDefinition {
+                        type_,
+                        items,
+                        attributes,
+                        inline_trailing_comments: inline_trailing_comments.clone(),
+                        following_comments: following_comments.clone(),
+                    }),
+                    location: ItemLocation::new(
+                        self.filename.clone(),
+                        Span::new(start_pos, end_pos),
+                    ),
+                })
+            }
+            TokenKind::Bitflags => {
+                self.advance();
+                let (name, _) = self.expect_ident()?;
+                self.expect(TokenKind::Colon)?;
+                let type_ = self.parse_type()?;
+                self.expect(TokenKind::LBrace)?;
+                let items = self.parse_bitflags_def_items()?;
+                self.expect(TokenKind::RBrace)?;
+
+                // Capture the end position
+                let end_pos = if self.pos > 0 {
+                    self.tokens[self.pos - 1].location.span.end
+                } else {
+                    self.current().location.span.end
+                };
+
+                Ok(ItemDefinition {
+                    visibility,
+                    name,
+                    doc_comments,
+                    inner: ItemDefinitionInner::Bitflags(BitflagsDefinition {
+                        type_,
+                        items,
+                        attributes,
+                        inline_trailing_comments,
+                        following_comments,
+                    }),
+                    location: ItemLocation::new(
+                        self.filename.clone(),
+                        Span::new(start_pos, end_pos),
+                    ),
+                })
+            }
+            _ => Err(ParseError::ExpectedItemDefinition {
+                found: self.peek().clone(),
+                location: self.current().location.clone(),
+            }),
+        }
+    }
+
+    pub(crate) fn parse_type_def_items(&mut self) -> Result<Vec<TypeDefItem>, ParseError> {
+        let mut items = Vec::new();
+
+        while !matches!(self.peek(), TokenKind::RBrace) {
+            // Collect non-doc comments (doc comments will be collected by parse_type_statement)
+            while matches!(
+                self.peek(),
+                TokenKind::Comment(_) | TokenKind::MultiLineComment(_)
+            ) {
+                if let Some(comment) = self.collect_comment() {
+                    items.push(TypeDefItem::Comment(comment));
+                }
+            }
+
+            if matches!(self.peek(), TokenKind::RBrace) {
+                break;
+            }
+
+            let mut stmt = self.parse_type_statement()?;
+            let statement_line = self.current().location.span.end.line;
+
+            // Optional trailing comma
+            if matches!(self.peek(), TokenKind::Comma) {
+                self.advance();
+            }
+
+            // Collect trailing comments after the comma, separating inline from following
+            while matches!(
+                self.peek(),
+                TokenKind::Comment(_) | TokenKind::MultiLineComment(_)
+            ) {
+                let comment_line = self.current().location.span.start.line;
+                if let Some(comment) = self.collect_comment() {
+                    if comment_line == statement_line {
+                        // Comment is on the same line as the field
+                        stmt.inline_trailing_comments.push(comment);
+                    } else {
+                        // Comment is on a following line
+                        stmt.following_comments.push(comment);
+                    }
+                }
+            }
+
+            items.push(TypeDefItem::Statement(stmt));
+        }
+
+        Ok(items)
+    }
+
+    pub(crate) fn parse_type_statement(&mut self) -> Result<TypeStatement, ParseError> {
+        let start_pos = self.current().location.span.start;
+
+        let doc_comments = self.collect_doc_comments();
+        let attributes = if matches!(self.peek(), TokenKind::Hash) {
+            self.parse_attributes()?
+        } else {
+            Attributes::default()
+        };
+
+        if matches!(self.peek(), TokenKind::Vftable) {
+            self.advance();
+            self.expect(TokenKind::LBrace)?;
+            let functions = self.parse_functions_in_block()?;
+            self.expect(TokenKind::RBrace)?;
+
+            let end_pos = if self.pos > 0 {
+                self.tokens[self.pos - 1].location.span.end
+            } else {
+                self.current().location.span.end
+            };
+
+            Ok(TypeStatement {
+                field: TypeField::Vftable(functions),
+                attributes,
+                doc_comments,
+                inline_trailing_comments: Vec::new(), // Will be populated by parse_type_def_items
+                following_comments: Vec::new(),
+                location: ItemLocation::new(self.filename.clone(), Span::new(start_pos, end_pos)),
+            })
+        } else {
+            let visibility = self.parse_visibility()?;
+            let (name, _) = self.expect_ident()?;
+            self.expect(TokenKind::Colon)?;
+            let type_ = self.parse_type()?;
+
+            let end_pos = if self.pos > 0 {
+                self.tokens[self.pos - 1].location.span.end
+            } else {
+                self.current().location.span.end
+            };
+
+            Ok(TypeStatement {
+                field: TypeField::Field(visibility, name, type_),
+                attributes,
+                doc_comments,
+                inline_trailing_comments: Vec::new(), // Will be populated by parse_type_def_items
+                following_comments: Vec::new(),
+                location: ItemLocation::new(self.filename.clone(), Span::new(start_pos, end_pos)),
+            })
+        }
+    }
+
+    pub(crate) fn parse_enum_def_items(&mut self) -> Result<Vec<EnumDefItem>, ParseError> {
+        let mut items = Vec::new();
+
+        while !matches!(self.peek(), TokenKind::RBrace) {
+            // Collect non-doc comments (doc comments will be collected by parse_enum_statement)
+            while matches!(
+                self.peek(),
+                TokenKind::Comment(_) | TokenKind::MultiLineComment(_)
+            ) {
+                if let Some(comment) = self.collect_comment() {
+                    items.push(EnumDefItem::Comment(comment));
+                }
+            }
+
+            if matches!(self.peek(), TokenKind::RBrace) {
+                break;
+            }
+
+            let mut stmt = self.parse_enum_statement()?;
+            let statement_line = self.current().location.span.end.line;
+
+            // Optional trailing comma
+            if matches!(self.peek(), TokenKind::Comma) {
+                self.advance();
+            }
+
+            // Collect trailing comments after the comma, separating inline from following
+            while matches!(
+                self.peek(),
+                TokenKind::Comment(_) | TokenKind::MultiLineComment(_)
+            ) {
+                let comment_line = self.current().location.span.start.line;
+                if let Some(comment) = self.collect_comment() {
+                    if comment_line == statement_line {
+                        // Comment is on the same line as the enum variant
+                        stmt.inline_trailing_comments.push(comment);
+                    } else {
+                        // Comment is on a following line
+                        stmt.following_comments.push(comment);
+                    }
+                }
+            }
+
+            items.push(EnumDefItem::Statement(stmt));
+        }
+
+        Ok(items)
+    }
+
+    pub(crate) fn parse_enum_statement(&mut self) -> Result<EnumStatement, ParseError> {
+        let start_pos = self.current().location.span.start;
+
+        let doc_comments = self.collect_doc_comments();
+        let attributes = if matches!(self.peek(), TokenKind::Hash) {
+            self.parse_attributes()?
+        } else {
+            Attributes::default()
+        };
+
+        let (name, _) = self.expect_ident()?;
+        let expr = if matches!(self.peek(), TokenKind::Eq) {
+            self.advance();
+            Some(self.parse_expr()?)
+        } else {
+            None
+        };
+
+        let end_pos = if self.pos > 0 {
+            self.tokens[self.pos - 1].location.span.end
+        } else {
+            self.current().location.span.end
+        };
+
+        Ok(EnumStatement {
+            name,
+            expr,
+            attributes,
+            doc_comments,
+            inline_trailing_comments: Vec::new(), // Will be populated by parse_enum_def_items
+            following_comments: Vec::new(),
+            location: ItemLocation::new(self.filename.clone(), Span::new(start_pos, end_pos)),
+        })
+    }
+
+    pub(crate) fn parse_bitflags_def_items(&mut self) -> Result<Vec<BitflagsDefItem>, ParseError> {
+        let mut items = Vec::new();
+
+        while !matches!(self.peek(), TokenKind::RBrace) {
+            // Collect non-doc comments (doc comments will be collected by parse_bitflags_statement)
+            while matches!(
+                self.peek(),
+                TokenKind::Comment(_) | TokenKind::MultiLineComment(_)
+            ) {
+                if let Some(comment) = self.collect_comment() {
+                    items.push(BitflagsDefItem::Comment(comment));
+                }
+            }
+
+            if matches!(self.peek(), TokenKind::RBrace) {
+                break;
+            }
+
+            let mut stmt = self.parse_bitflags_statement()?;
+            let statement_line = self.current().location.span.end.line;
+
+            // Optional trailing comma
+            if matches!(self.peek(), TokenKind::Comma) {
+                self.advance();
+            }
+
+            // Collect trailing comments after the comma, separating inline from following
+            while matches!(
+                self.peek(),
+                TokenKind::Comment(_) | TokenKind::MultiLineComment(_)
+            ) {
+                let comment_line = self.current().location.span.start.line;
+                if let Some(comment) = self.collect_comment() {
+                    if comment_line == statement_line {
+                        // Comment is on the same line as the bitflag
+                        stmt.inline_trailing_comments.push(comment);
+                    } else {
+                        // Comment is on a following line
+                        stmt.following_comments.push(comment);
+                    }
+                }
+            }
+
+            items.push(BitflagsDefItem::Statement(stmt));
+        }
+
+        Ok(items)
+    }
+
+    pub(crate) fn parse_bitflags_statement(&mut self) -> Result<BitflagsStatement, ParseError> {
+        let start_pos = self.current().location.span.start;
+
+        let doc_comments = self.collect_doc_comments();
+        let attributes = if matches!(self.peek(), TokenKind::Hash) {
+            self.parse_attributes()?
+        } else {
+            Attributes::default()
+        };
+
+        let (name, _) = self.expect_ident()?;
+        self.expect(TokenKind::Eq)?;
+        let expr = self.parse_expr()?;
+
+        let end_pos = if self.pos > 0 {
+            self.tokens[self.pos - 1].location.span.end
+        } else {
+            self.current().location.span.end
+        };
+
+        Ok(BitflagsStatement {
+            name,
+            expr,
+            attributes,
+            doc_comments,
+            inline_trailing_comments: Vec::new(), // Will be populated by parse_bitflags_def_items
+            following_comments: Vec::new(),
+            location: ItemLocation::new(self.filename.clone(), Span::new(start_pos, end_pos)),
+        })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::{
+        grammar::{
+            IntFormat,
+            test_aliases::{int_literal, int_literal_with_format, *},
+        },
+        parser::parse_str_for_tests,
+        span::StripLocations,
+    };
+    use pretty_assertions::assert_eq;
+
+    #[test]
+    fn can_parse_enum() {
+        let text = r#"
+        #[singleton(0x1234)]
+        pub enum TestType: u32 {
+            Item0 = -5,
+            #[default]
+            Item1,
+            Item2,
+            Item3 = 10,
+            Item4
+        }
+        "#;
+
+        let ast = M::new().with_definitions([ID::new(
+            (V::Public, "TestType"),
+            ED::new(
+                T::ident("u32"),
+                [
+                    ES::field_with_expr("Item0", int_literal(-5)),
+                    ES::field("Item1").with_attributes([A::default()]),
+                    ES::field("Item2"),
+                    ES::field_with_expr("Item3", int_literal(10)),
+                    ES::field("Item4"),
+                ],
+                [A::singleton(0x1234)],
+            ),
+        )]);
+
+        assert_eq!(parse_str_for_tests(text).unwrap().strip_locations(), ast);
+    }
+
+    #[test]
+    fn can_parse_bitflags() {
+        let text = r#"
+        #[singleton(0x1234)]
+        pub bitflags TestType: u32 {
+            #[default]
+            Item1 = 0b0001,
+            Item2 = 0b0010,
+            Item3 = 0b0100,
+            Item4 = 0b1000,
+        }
+        "#;
+
+        let ast = M::new().with_definitions([ID::new(
+            (V::Public, "TestType"),
+            BFD::new(
+                T::ident("u32"),
+                [
+                    BFS::field("Item1", int_literal_with_format(1, IntFormat::Binary))
+                        .with_attributes([A::default()]),
+                    BFS::field("Item2", int_literal_with_format(2, IntFormat::Binary)),
+                    BFS::field("Item3", int_literal_with_format(4, IntFormat::Binary)),
+                    BFS::field("Item4", int_literal_with_format(8, IntFormat::Binary)),
+                ],
+                [A::singleton(0x1234)],
+            ),
+        )]);
+
+        assert_eq!(parse_str_for_tests(text).unwrap().strip_locations(), ast);
+    }
+}
